added functionality to add projects
