+++ src/js/index.js
@@ -1,19 +1,10 @@
 import '../css/style.scss';
 import Project from './model/project';
-import { addProject, updateLocalStorage, deleteProject } from './controller/projectCtrl';
+import { addProject, updateProjects, getProjectArr, updateLocalStorage, deleteProject } from './controller/projectCtrl';
 import TodoItem from './model/todo';
 import renderProject from './view/projectView';
 import * as todoListCtrl from './controller/todoListCtrl';
 import dom from './view/domStrings';
-
-let projects = [];
-export default function getProjectArr() {
-  const localStorageArr = JSON.parse(localStorage.getItem('projects'));
-  if (localStorageArr) {
-    projects = [...localStorageArr];
-  }
-  return projects;
-}
 
 function createProject(title) {
   const p = Project(title);
@@ -34,5 +25,6 @@ const mainController = (() => {
   };
   return { eventHandler };
 })();
+updateProjects();
 renderProject();
 mainController.eventHandler();
